Extract `Fail` from `Json`
