+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/http/ClientSpec.scala
@@ -22,8 +22,8 @@ import cats.implicits._
 
 import sbt.util.Logger
 
+import com.alejandrohdezma.sbt.github.failure.Fail
 import com.alejandrohdezma.sbt.github.json.Decoder
-import com.alejandrohdezma.sbt.github.json.Json.Fail
 import com.alejandrohdezma.sbt.github.syntax.json.JsonValueOps
 import com.alejandrohdezma.sbt.github.withServer
 import org.http4s.dsl.io._
