+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/syntax/JsonSyntaxSpec.scala
@@ -16,9 +16,9 @@
 
 package com.alejandrohdezma.sbt.github.syntax
 
+import com.alejandrohdezma.sbt.github.failure.Fail
+import com.alejandrohdezma.sbt.github.failure.Fail._
 import com.alejandrohdezma.sbt.github.json.Json
-import com.alejandrohdezma.sbt.github.json.Json.Fail
-import com.alejandrohdezma.sbt.github.json.Json.Fail._
 import com.alejandrohdezma.sbt.github.syntax.json._
 import org.specs2.mutable.Specification
